Extract the scan sharding, progress and task-join helpers into utils.
